--- examples/with-netlify/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "ingest-with-netlify",
-  "version": "1.0.0",
-  "description": "A simple boilerplate for using Ingest with Netlify.",
-  "private": true,
-  "scripts": {
-    "generate": "yarn build:tsc && yarn build:files",
-    "build:tsc": "tsc",
-    "build:files": "ts-node src/build.ts",
-    "dev": "ts-node src/server.ts"
-  },
-  "dependencies": {
-    "@stackpress/ingest-netlify": "0.3.2"
-  },
-  "devDependencies": {
-    "@types/node": "22.9.3",
-    "ts-node": "10.9.2",
-    "typescript": "5.7.2"
-  }
-}
