--- examples/with-netlify/src/build.ts
@@ -1,4 +1,0 @@
-import server from './routes';
-server.build().then(({ build }) => {
-  console.log('done.', build);
-});
