--- examples/with-netlify/src/routes/home.ts
@@ -1,5 +1,0 @@
-import { Context, Response } from '@stackpress/ingest';
-
-export default async function HomePage(req: Context, res: Response) { 
-  res.setHTML('hello');
-};
