--- examples/with-netlify/src/routes/redirect.ts
@@ -1,5 +1,0 @@
-import { Context, Response } from '@stackpress/ingest';
-
-export default function Redirect(req: Context, res: Response) {
-  res.redirect('/user');
-};
