--- examples/with-vercel/plugin.d.ts
@@ -1,2 +1,0 @@
-import plugin from './dist/plugin';
-export default plugin;
