--- packages/ingest-netlify/README.md
@@ -1,3 +1,0 @@
-# ᗊ Ingest - Netlify
-
-An unopinionated, event driven, pluggable, Netlify framework.
