--- packages/ingest-netlify/index.d.ts
@@ -1,3 +1,0 @@
-import vercel from './dist';
-export * from './dist';
-export default vercel;
