--- packages/ingest-vercel/README.md
@@ -1,3 +1,0 @@
-# ᗊ Ingest - Vercel
-
-An unopinionated, event driven, pluggable, Vercel framework.
