--- packages/ingest-vercel/index.js
@@ -1,3 +1,0 @@
-const vercel = require('./dist');
-module.exports = vercel.default
-Object.assign(module.exports, vercel);
