--- packages/ingest-vercel/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "@stackpress/ingest-vercel",
-  "version": "0.3.2",
-  "license": "Apache-2.0",
-  "description": "Using Ingest to generate Vercel functions.",
-  "author": "Chris <chris@stackpress.io>",
-  "homepage": "https://github.com/stackpress/ingest/tree/main/packages/types",
-  "bugs": "https://github.com/stackpress/ingest/issues",
-  "repository": "stackpress/ingest",
-  "main": "index.js",
-  "types": "index.d.ts",
-  "files": [
-    "index.js",
-    "index.d.ts",
-    "dist",
-    "LICENSE",
-    "README.md"
-  ],
-  "scripts": {
-    "build": "tsc"
-  },
-  "dependencies": {
-    "@stackpress/ingest": "0.3.2"
-  },
-  "devDependencies": {
-    "@types/node": "22.9.3",
-    "ts-node": "10.9.2",
-    "typescript": "5.7.2"
-  }
-}
