+++ packages/ingest/src/index.ts
@@ -6,7 +6,7 @@ import { ConfigLoader, PluginLoader } from './Loader';
 import Request from './Request';
 import Response from './Response';
 import Router, { ServerRouter } from './Router';
-import Server from './Server';
+import Server, { handler, gateway } from './Server';
 import { ReadSession, WriteSession, session } from './Session';
 
 export type * from './types';
@@ -15,6 +15,8 @@ export * from './helpers';
 export {
   cookie,
   session,
+  handler, 
+  gateway,
   Exception,
   ConfigLoader, 
   PluginLoader,
